Add dns:encode_message/2
Adds options for truncating a message or spreading it over multiple
messages at encode time. The former is the default, the later can be
enabled to suit AXFRs or LLQs. If a maximum message size isn't specified
it is instead taken from the messages optrr or failing that falls back
to 512. If TSIG credentials are supplied the message will be truncated
to fit the TSIG RR if need be.