Minor optimisations (#65)
* Use maps instead of dicts and gb_trees in dnssec

dnssec code now uses only maps instead of gb_trees and dicts. Note in particular that dicts have very very poor performance and are strongly unrecommended.

* Optimise nsec3 iterative hashing function

Note however that this is an expensive operation on the beam due to
immutability, if iterations are too high and/or this operation happens
too often, we should consider making it a NIF like it is done at
https://github.com/esl/fast_pbkdf2